// frontend/src/modules/3DViewerNew/customDeckGlLayers/SeismicFenceMeshLayer copy/_private/webworker/types.ts
export type WebworkerParameters = {
    verticesArray: Float32Array;
    indicesArray: Uint32Array;
    numSamplesU: number;
    numSamplesV: number;
    boundingBox: number[][];
    zIncreasingDownwards: boolean;
};

export type WebworkerResult = {
    verticesArray: Float32Array;
    indicesArray: Uint32Array;
};